requirements.txt
polars>=1.0
pytest
pyright
rich
msgpack
pyzmq

feedparser
psutil

python-dotenv
fastapi>=0.115
uvicorn[standard]>=0.30
aiohttp
requests
